+++ internal/nginx/pid_unix.go
@@ -2,7 +2,10 @@
 
 package nginx
 
-import "syscall"
+import (
+	"os"
+	"syscall"
+)
 
 func reloadProcess(pid int) error {
 	return syscall.Kill(pid, syscall.SIGHUP)
@@ -10,4 +13,12 @@ func reloadProcess(pid int) error {
 
 func terminateProcess(pid int, sig syscall.Signal) error {
 	return syscall.Kill(pid, sig)
+}
+
+func isPIDAlive(pid int) bool {
+	proc, err := os.FindProcess(pid)
+	if err != nil {
+		return false
+	}
+	return proc.Signal(syscall.Signal(0)) == nil
 }